code format

+++ wm800/board/platform/arch/xt804/libc/liteos/math/copysign.c
@@ -15,7 +15,7 @@
 
 #include "libm.h"
 
-double copysign(double x, double y)
+double copysign(long double x, long double y)
 {
     union {
         double f;
